Removendo float desnecessário

+++ lista_exec7.py
@@ -1,5 +1,5 @@
 celcius = float(input("Digite a temperatura em Celcius: "))
 
-fahrenheit = float(celcius*1.8+32)
+fahrenheit = (celcius*1.8+32)
 
 print(f"A temperatura em Fahrenheit é {fahrenheit}")
